requirements.txt
fastapi
uvicorn
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
psycopg2-binary
pydantic[email]
python-dotenv
email-validator
alembic
psycopg2-binary

